refactoring "routes" name into "router"
